Add support for garage gate
I own Krispol Starcus garage gate drive with a "Gate Pro Connector"
which turnes out to be compatible with this library after a few small
changes.

Tested this on the 2 devices that I own and works without any
issues. I'm able to open, close and set position of the gate.

Example status:
<MotionBlind mac: c8c9a305eafb, type: RollerBlind, status: Stopped, position: 100 %, angle: 0.0, limit: BothLimitsDetected, battery: None, None %, None V, charging: None, RSSI: -53 dBm, com: None>

Signed-off-by: Krzysztof Opasiak <krzysztof.opasiak@neat.no>