Allow gateway and blind to have the same mac address
In the Krispol gate drive the gateway and the blind is reported using
the same mac address. This isn't an issue for synchronous API but when
using the asyncio the blind callback was never triggered.

This commit makes us process both gateway and blind callbacks if they
are reported with the same MAC address.

Signed-off-by: Krzysztof Opasiak <krzysztof.opasiak@neat.no>